requirements.txt
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0
Pillow>=9.5.0
requests>=2.28.2
pyinstaller>=6.11.1